% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchordist_23.R
\name{anchordist}
\alias{anchordist}
\title{Compute cosine distances anchored to first n lemmas}
\usage{
anchordist(targetdf, anchorsize, windowsize)
}
\arguments{
\item{targetdf}{a dataframe with long list of target words}

\item{anchorsize}{the size of the anchor block}

\item{windowsize}{the size of each window to compute cosine distances from anchor block}
}
\description{
Details here
}
