% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_allsteps.R
\name{clean_df}
\alias{clean_df}
\title{Clean Raw Text}
\usage{
clean_df(x)
}
\arguments{
\item{x}{A dataframe with raw text in a column titled "doc_text".}
}
\value{
the function \code{clean_df} with return a dataframe of cleaned text documents, their document ID and the raw text.
}
\description{
Use helper functions to clean raw text
}
