% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_df_bygroup.R
\name{clean_df_bygroup}
\alias{clean_df_bygroup}
\title{Applies cleaning function to multiple texts}
\usage{
clean_df_bygroup(datafile)
}
\arguments{
\item{x}{A dataframe with of mutiple documents denoted in a column titled "doc_id" and raw text in a column titled "doc_text" .}
}
\value{
the function \code{clean_df_bygroup} with return a dataframe of cleaned text documents, their document ID and the raw text.
}
\description{
Uses helper functions to clean raw text
}
