% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_df_nopronouns.R
\name{clean_df_nopronouns}
\alias{clean_df_nopronouns}
\title{Clean Raw Text keeping pronouns}
\usage{
clean_df_nopronouns(x)
}
\arguments{
\item{x}{A dataframe with raw text in a column titled "doc_text".}
}
\value{
the function \code{clean_df_nopronouns} with return a dataframe of cleaned text documents, their document ID and the raw text.
}
\description{
Use helper functions to clean raw text
}
