% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanme_23.R
\name{cleanme}
\alias{cleanme}
\title{Clean Raw Text}
\usage{
cleanme(x)
}
\arguments{
\item{dataframe}{with raw text in a column titled "doc_text". Usually first generated by readme()}
}
\value{
a dataframe of cleaned text documents (doc_clean) their document ID and raw text preserved.
}
\description{
Clean and format raw text
}
\keyword{internal}
