% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pronouns_titles.R
\name{pronouns_titles}
\alias{pronouns_titles}
\title{Replace pronouns and titles with XX}
\usage{
pronouns_titles(x)
}
\arguments{
\item{x}{A vector of words.}
}
\description{
this code uses gsub to remove titles (Mr., Ms.) and pronouns (him, her) and replaces them with the XX (person, man, women).
}
\keyword{internal}
