% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readin.R
\name{readin}
\alias{readin}
\title{Read in Text}
\arguments{
\item{folder_name}{A string.}
}
\value{
the function \code{readin} with return a single dataframe of text documents and their document ID
}
\description{
Read in .txt documents from a specified folder found in the current working directory.
}
\examples{
readin("text")
readin("data")
}
