% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readme_23.R
\name{readme}
\alias{readme}
\title{Read in Text}
\usage{
readme(folder_name = "my_texts")
}
\arguments{
\item{folder_name}{A string.}
}
\value{
the function \code{readme} with return a single dataframe of text documents and their document ID
}
\description{
Read in .txt documents from a specified folder found in the current working directory.
}
\examples{
readtxt("text")
readtxt("data")
}
