% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readtxt.R
\name{readtxt}
\alias{readtxt}
\title{Read in Text}
\usage{
readtxt(folder_name = "data")
}
\arguments{
\item{folder_name}{A string.}
}
\value{
the function \code{readtxt} with return a single dataframe of text documents and their document ID
}
\description{
Read in .txt documents from a specified folder found in the current working directory.
}
\examples{
readtxt("text")
readtxt("data")
}
