% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowwise_cosine_simil.R
\name{rowwise_cosine_simil}
\alias{rowwise_cosine_simil}
\title{Import dplyr}
\usage{
rowwise_cosine_simil(
  data_file = x,
  word_rating = wordvec,
  colname1 = "word",
  colname2 = "word"
)
}
\description{
Import dplyr
}
