% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowwise_euc_diff.R
\name{rowwise_euc_diff}
\alias{rowwise_euc_diff}
\title{Rowwise Euclidean Distance}
\arguments{
\item{targetdf}{a data frame with long list of target words.}

\item{lookupdb}{a data frame containing ratings}

\item{colname1}{the column name from targetdf used to join}

\item{colname2}{the column name from lookupdb used to join}
}
\description{
Rowwise Euclidean Distance
}
