% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanme_23.R
\name{cleanme}
\alias{cleanme}
\title{Clean Raw Text}
\usage{
cleanme(dat)
}
\arguments{
\item{x}{A dataframe with raw text in a column titled "doc_text".}
}
\value{
the function \code{cleanme} with return a dataframe of cleaned text documents, their document ID and the raw text.
}
\description{
Use helper function to clean raw text
}
\keyword{internal}
