% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_df_bygroup.R
\name{cleanme}
\alias{cleanme}
\title{Applies cleaning function to multiple texts}
\usage{
cleanme(datafile, replace_pronoun = FALSE)
}
\arguments{
\item{x}{A dataframe with of mutiple documents denoted in a column titled "doc_id" and raw text in a column titled "doc_text" .}
}
\value{
the function \code{cleanme} with return a dataframe of cleaned text documents, their document ID and the raw text.
}
\description{
Uses helper functions to clean raw text
}
