% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distme_23.R
\name{distme}
\alias{distme}
\title{Rowwise Cosine Distance Calculations}
\usage{
distme(targetdf, lemmatize = TRUE)
}
\arguments{
\item{targetdf}{a dataframe with long list of target words}
}
\description{
Details here
}
