% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readin.R
\name{readin}
\alias{readin}
\title{Read in Text}
\usage{
readin(folder_name = "data")
}
\arguments{
\item{folder_name}{A string.}

\item{y}{A number.}
}
\value{
Dataframe of text documents and their document ID
}
\description{
Read in .txt documents from a specified folder in the working directory
}
\examples{
readin("text")
readin("data")
}
