% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readtxt.R
\name{readtxt}
\alias{readtxt}
\title{Read in Text}
\usage{
readtxt()
}
\value{
the function \code{readtxt} with return a single dataframe of text documents and their document ID
}
\description{
Read in .txt documents from a specified folder found in the current working directory.
}
