% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowwise_cosine_simil.R
\name{rowwise_cosine_simil}
\alias{rowwise_cosine_simil}
\title{Rowwise Cosine Calculations}
\usage{
rowwise_cosine_simil(
  targetdf = x,
  lookupdb = wordvec,
  colname1 = NULL,
  colname2 = NULL
)
}
\arguments{
\item{targetdf}{a data frame with long list of target words.}

\item{lookupdb}{a data frame containing ratings}

\item{colname1}{the column name from targetdf used to join}

\item{colname2}{the column name from lookupdb used to join}
}
\description{
Details here
}
