% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowwise_euc_diff.R
\name{rowwise_euc_diff}
\alias{rowwise_euc_diff}
\title{Import dplyr}
\usage{
rowwise_euc_diff(
  data_file = x,
  word_rating = wordvec,
  colname1 = word,
  colname2 = word
)
}
\description{
Import dplyr
}
